// src/modules/mgmt-api-client/model/contractNegotiationDto.ts
/**
 * EDC REST API
 * EDC REST APIs - merged by OpenApiMerger
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ContractNegotiationDto { 
    contractAgreementId?: string;
    counterPartyAddress?: string;
    createdAt?: number;
    errorDetail?: string;
    id?: string;
    protocol?: string;
    state?: string;
    type?: ContractNegotiationDto.TypeEnum;
    updatedAt?: number;
}
export namespace ContractNegotiationDto {
    export type TypeEnum = 'CONSUMER' | 'PROVIDER';
    export const TypeEnum = {
        Consumer: 'CONSUMER' as TypeEnum,
        Provider: 'PROVIDER' as TypeEnum
    };
}