// src/modules/mgmt-api-client/model/negotiationInitiateRequestDto.ts
/**
 * EDC REST API
 * EDC REST APIs - merged by OpenApiMerger
 *
 * The version of the OpenAPI document: 0.0.1-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ContractOfferDescription } from './contractOfferDescription';


export interface NegotiationInitiateRequestDto { 
    connectorAddress: string;
    connectorId: string;
    offer: ContractOfferDescription;
    protocol: string;
}